Broken PRD function.
